Refactor PowerNode to move spec fields to status
PowerNode is auto-created and managed by the controller - users don't
create or configure it. All the data should be stored in the status
field , as it represents the observed state of the system, rather than
desired configuration.

Remove unused types defined in the powernode API.
Remove redundant field - nodeName as the object name already represents
the node name.